adds , , renames folders
