+++ project/src/components/header/header.tsx
@@ -1,13 +1,11 @@
+import { Logo } from 'components';
+
 export default function Header() {
   return (
     <header className="header">
       <div className="container">
         <div className="header__wrapper">
-          <div className="header__left">
-            <a className="header__logo-link header__logo-link--active">
-              <img className="header__logo" src="img/logo.svg" alt="six cities logo" width="81" height="41" />
-            </a>
-          </div>
+          <Logo />
           <nav className="header__nav">
             <ul className="header__nav-list">
               <li className="header__nav-item user">
