+++ project/src/components/index.ts
@@ -1,3 +1,7 @@
 export { default as App } from './app/app';
 export { default as CitiesCard } from './cities-card/cities-card';
 export { default as Header } from './header/header';
+export { default as Layout } from './layout/layout';
+export { default as StarIcon } from './star-icon/star-icon';
+export { default as Logo } from './logo/logo';
+
