+++ project/src/components/layout/layout.tsx
@@ -1,0 +1,12 @@
+import { Header } from 'components';
+import { Fragment } from 'react';
+import { Outlet } from 'react-router-dom';
+
+export default function Layout() {
+  return (
+    <Fragment>
+      <Header />
+      <Outlet />
+    </Fragment>
+  );
+}
