+++ project/src/const.tsx
@@ -1,5 +1,5 @@
 export enum AppRoute {
   Main = '/',
   Login = '/login',
-  Room = '/offer/:id',
+  Offer = '/offer/:id',
 }
