--- project/src/pages/404-screen/404-screen.tsx
@@ -1,10 +1,0 @@
-import { Header } from 'components';
-
-export default function Screen404() {
-  return (
-    <>
-      <Header />
-      <h1 className='property__name'>Ошибка 404. Страницы не существует.</h1>
-    </>
-  );
-}
