+++ project/src/pages/index.ts
@@ -1,4 +1,4 @@
 export { default as CitiesHomeScreen } from './cities-home-screen/cities-home-screen';
 export { default as LoginScreen } from './login-screen/login-screen';
-export { default as PropertyScreen } from './property-screen/property-screen';
-export { default as Screen404 } from './404-screen/404-screen';
+export { default as OfferScreen } from './offer-screen/offer-screen';
+export { default as Screen404 } from './screen-404/screen-404';
