+++ project/src/pages/screen-404/screen-404.tsx
@@ -1,0 +1,5 @@
+export default function Screen404() {
+  return (
+    <h1 className='property__name'>Ошибка 404. Страницы не существует.</h1>
+  );
+}
